// src/gen/bikehub/model/paymentMethod.ts
/**
 * BikeHubParthenope
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PaymentType } from './paymentType';


export interface PaymentMethod {
    id?: string;
    type?: PaymentType;
    cc?: string;
    cvc?: string;
    holder?: string;
    expireAt?: string;
}