// src/gen/bikehub/model/updatePaymentMethod.ts
/**
 * BikeHubParthenope
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PaymentType } from './paymentType';


export interface UpdatePaymentMethod {
    type?: PaymentType;
    cc?: string;
    cvc?: string;
    holder?: string;
    expireAt?: string;
}